+++ CHANGELOG.md
@@ -3,9 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-Dates are in ISO-8601 standart (trying to be)
+Dates are in ISO-8601 standart (trying to be) (YYYY-MM-DD)
 
-## [Unreleased]
+## [1.0.5] - 2023-06-12
+### Added
+- `IReturnableToPool` for `OnReturnToPool()` callback
+- `ISetupable` with `PoolableItemSetupParameters` for
+### Changed
+- `ObjectPooler`'s `GetFreeObject()` now handles auto setup with `ISetupable` and `PoolableItemSetupParameters`
+- `ComponentObjectPooler`'s `ReturnToPool()` now calls `IReturnableToPool` back on return
 
 ## [1.0.4] - 2022-08-28
 ### Added